requirements.txt
fastapi~=0.115.12
sqlalchemy~=2.0.40
postgres
psycopg2
pytest~=8.3.5
uvicorn~=0.34.1
dotenv~=0.9.9
bcrypt~=4.3.0
pyjwt[crypto]~=2.10.1
cryptography~=44.0.2
jose~=1.0.0
pydantic~=2.11.3
python-dotenv~=1.1.0
alembic~=1.15.2
pydantic[email]
python-multipart
httpx~=0.28.1
starlette~=0.46.2